requirements.txt
browser-use==0.5.6
python-dotenv==1.1.1
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
